+++ CHANGELOG.md
@@ -94,6 +94,7 @@ in a future release.
     - `SqlServerEndpointPermission` was renamed to `SqlEndpointPermission`.
     - `SqlServerLogin` was renamed to `SqlLogin`.
     - `SqlServerMaxDop` was renamed to `SqlMaxDop`.
+    - `SqlServerMemory` was renamed to `SqlMemory`.
   - Changed all resource prefixes from `MSFT_` to `DSC_` ([issue #1496](https://github.com/dsccommunity/SqlServerDsc/issues/1496)).
     _Deprecated resource has not changed prefix._
   - All resources are now using the common module DscResource.Common.